Hopefully made it work on server now.
Don't check for __GHCJS__ in Setup.hs, nix builds it with ghc!

+++ front/preBuild.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 dev=$1
 distPath=dist/build/gonimo-front/gonimo-front.jsexe
-rm -Rf ${distPath}/*
+[ -d ${distPath} ] && rm -Rf ${distPath}/*
